fixes potential problem with petrol consumption
The previous choice only consumed petrol (`tick`) when a reduction was
performed, which meant that arbitrarily large data in normal form would
consume no petrol, because they had no reductions. This would be a
possible vector of attack, because the traversal over such structures
is linear time in the size of the structure, but costs nothing,
allowing arbitrarily long run times for free.

The new version consumes one unit of petrol every time the evaluation
function (more specifically, the auxiliary function `go`) is entered,
so that any and all computation costs, even the trivial kind.